Code according UnitTests about Review and Comment models
