+++ api_yamdb/api/management/commands/fulfill.py
@@ -4,8 +4,9 @@ from django.core.management.base import BaseCommand
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import IntegrityError
 
-from titles.models import (Title, GenreTitle, Review, Comment,
-                           Genre, User, Category)
+from titles.models import Title, GenreTitle, Genre, Category
+from reviews.models import Review, Comment
+from users.models import User
 
 
 class Command(BaseCommand):
