+++ api_yamdb/reviews/admin.py
@@ -1,0 +1,19 @@
+from django.contrib import admin
+
+from reviews.models import Review, Comment
+
+
+@admin.register(Comment)
+class CommentAdmin(admin.ModelAdmin):
+    list_display = ('review', 'text', 'author', 'pub_date',)
+    search_fields = ('review',)
+    list_filter = ('review',)
+    empty_value_display = '-пусто-'
+
+
+@admin.register(Review)
+class ReviewAdmin(admin.ModelAdmin):
+    list_display = ('title', 'text', 'author', 'score',)
+    search_fields = ('pub_date',)
+    list_filter = ('pub_date',)
+    empty_value_display = '-пусто-'
