+++ api_yamdb/titles/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Genre, Category, Title, Comment, Review
+from .models import Genre, Category, Title
 
 
 @admin.register(Title)
@@ -24,19 +24,3 @@ class CategoryAdmin(admin.ModelAdmin):
     """Настройка панели администратора для модели Category(Категории)"""
     list_display = ('name', 'slug')
     prepopulated_fields = {'slug': ('name',)}
-
-
-@admin.register(Comment)
-class CommentAdmin(admin.ModelAdmin):
-    list_display = ('review', 'text', 'author', 'pub_date',)
-    search_fields = ('review',)
-    list_filter = ('review',)
-    empty_value_display = '-пусто-'
-
-
-@admin.register(Review)
-class ReviewAdmin(admin.ModelAdmin):
-    list_display = ('title', 'text', 'author', 'score',)
-    search_fields = ('pub_date',)
-    list_filter = ('pub_date',)
-    empty_value_display = '-пусто-'
