+++ api_yamdb/titles/models.py
@@ -1,7 +1,4 @@
 from django.db import models
-from django.core.validators import MaxValueValidator, MinValueValidator
-
-from users.models import User
 
 
 class Title(models.Model):
@@ -88,73 +85,3 @@ class Category(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class Review(models.Model):
-    """Отзывы на произведения. Отзыв привязан к определённому произведению."""
-    title = models.ForeignKey(
-        Title,
-        null=True,
-        on_delete=models.CASCADE,
-        verbose_name='Произведение',
-    )
-    text = models.TextField(verbose_name='Текст отзыва')
-    author = models.ForeignKey(
-        User,
-        on_delete=models.CASCADE,
-        verbose_name='Автор',
-    )
-    score = models.PositiveSmallIntegerField(
-        verbose_name='Рейтинг',
-        validators=[
-            MinValueValidator(1, 'Допустимы значения от 1 до 10'),
-            MaxValueValidator(10, 'Допустимы значения от 1 до 10')
-        ]
-    )
-    pub_date = models.DateTimeField(
-        auto_now_add=True,
-        verbose_name='Дата добавления'
-    )
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=['title', 'author'],
-                name='unique_author_review'
-            )
-        ]
-        ordering = ('-pub_date',)
-        verbose_name = 'Отзыв'
-        verbose_name_plural = 'Отзывы'
-        default_related_name = 'reviews'
-
-    def __str__(self):
-        return f'{self.text}'
-
-
-class Comment(models.Model):
-    """Комментарии к отзывам. Комментарий привязан к определённому отзыву."""
-    review = models.ForeignKey(
-        Review,
-        on_delete=models.CASCADE,
-        verbose_name='Отзыв',
-    )
-    text = models.TextField(verbose_name='Комментарий')
-    author = models.ForeignKey(
-        User,
-        on_delete=models.CASCADE,
-        verbose_name='Автор',
-    )
-    pub_date = models.DateTimeField(
-        auto_now_add=True,
-        verbose_name='Дата добавления'
-    )
-
-    class Meta:
-        verbose_name = 'Комментарий'
-        verbose_name_plural = 'Комментарии'
-        ordering = ['pub_date']
-        default_related_name = 'comments'
-
-    def __str__(self):
-        return f'{self.text}'
