+++ requirements.txt
@@ -7,4 +7,4 @@ pytest-django==4.4.0
 pytest-pythonpath==0.7.3
 python-dotenv==1.0.0
 djangorestframework-simplejwt==4.7.2
-django-filter-22.1
+django-filter==22.1
